building deleteOne user functionality

+++ users/users-dao.js
@@ -6,14 +6,12 @@ export const updateUser = async (uid, userUpdate) => await usersModel.updateOne(
     {_id: uid},
     {$set: userUpdate}
 );
-export const deleteUser = async (uid) => await usersModel.deleteOne(
-    {_id: uid}
-);
+export const deleteUser = async (uid) => await usersModel.deleteOne({ _id: uid });
 
 export const findAllUsers = async () => await usersModel.find();
 
-export const findUserById = async (uid) => await usersModel.findById(uid, {password: false});
-//export const findUserById = async (uid) => await usersModel.findById(uid);
+//export const findUserById = async (uid) => await usersModel.findById(uid, {password: false});
+export const findUserById = async (uid) => await usersModel.findById(uid);
 
 export const findUserByUsername = async (username) => await usersModel.findOne(
     {username: username}, {password: false}
